requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
audioop-lts==0.2.2
blinker==1.9.0
certifi==2025.8.3
click==8.2.1
discord.py==2.5.2
Flask==3.1.1
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.6.3
passlib==1.7.4
bcrypt==3.2.2
ollama==0.5.3
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
pyrefly==0.25.1
python-dotenv==1.1.1
setuptools==78.1.1
sniffio==1.3.1
typing-inspection==0.4.1
typing_extensions==4.15.0
watchfiles==1.1.0
Werkzeug==3.1.3
wheel==0.45.1
yarl==1.20.1
fastapi==0.115.5
uvicorn==0.32.1
redis==5.2.1
Pillow==12.0.0